backend/requirements.txt
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
requests==2.32.3
fastapi==0.115.6
uvicorn==0.34.0
